add some translations and split code
