+++ apps/web/src/lib/l10n/en.json
@@ -17,8 +17,10 @@
   "links.about": "About Us",
   "links.events": "Events",
   "links.contact": "Contact",
-
+  
+  "readMore": "Read more",
   "donate": "Donate",
+  "volunteer": "Volunteer",
   "followUs": "Follow us",
   "copyright": "Copyright © 2014, All Right Reserved Stichting Oekraïners in Nederland"
 }
